+++ Semana_3/gobarber-native/gobarberapp/src/routes/index.tsx
@@ -1,19 +1,21 @@
-import React from 'react'
-import {createStackNavigator} from '@react-navigation/stack'
-import SignIn from '../pages/SignIn'
-import SignUp from '../pages/SignUp'
+import React from "react";
+import { createStackNavigator } from "@react-navigation/stack";
+import SignIn from "../pages/SignIn";
+import SignUp from "../pages/SignUp";
 
 const Auth = createStackNavigator();
 
+const AuthRoutes: React.FC = () => (
+  <Auth.Navigator
+    screenOptions={{
+      headerShown: false,
+      cardStyle: { backgroundColor: "#312e38" },
+    }}
+    //initialRouteName="SignUp" Define a rota Inicial, do contrario ele segue a ordem
+  >
+    <Auth.Screen name="SignIn" component={SignIn} />
+    <Auth.Screen name="SignUp" component={SignUp} />
+  </Auth.Navigator>
+);
 
-const AuthRoutes : React.FC = () =>(
-    <Auth.Navigator screenOptions ={{
-        headerShown: false,
-        cardStyle: {backgroundColor: "#312e38"}
-    }}>
-        <Auth.Screen name='SignIn' component={SignIn}/>
-        <Auth.Screen name='SignUp' component={SignUp}/>
-    </Auth.Navigator>
-)
-
-export default AuthRoutes
+export default AuthRoutes;